feature: added go.mod file and helper.go file | Implemented map function to hold user data
