+++ go.mod
@@ -1,0 +1,3 @@
+module golang-ticket-sales-app
+
+go 1.13
